实现facet相关操作

+++ pshow_repository/src/main/java/org/pshow/repo/datamodel/content/PropertyValue.java
@@ -45,7 +45,7 @@ public class PropertyValue {
                 this.value = pm.getStringValue();
                 break;
             case INT:
-                this.value = pm.getLongValue();
+                this.value = pm.getLongValue().intValue();
                 break;
             case LONG:
                 this.value = pm.getLongValue();
